+++ generator/src/loaders.rs
@@ -69,6 +69,7 @@ impl Function {
             TypeName::device(),
             TypeName::command_buffer(),
             TypeName::queue(),
+            TypeName::render_pass(),
         ];
 
         let instance_types = [TypeName::instance(), TypeName::physical_device()];