+++ generator/src/name.rs
@@ -191,6 +191,10 @@ impl TypeName {
         TypeName::new("VkQueue")
     }
 
+    pub fn render_pass() -> TypeName {
+        TypeName::new("VkRenderPass")
+    }
+
     pub fn physical_device() -> TypeName {
         TypeName::new("VkPhysicalDevice")
     }